*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk@1199 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ ac/source/src/rendertext.cpp
@@ -134,7 +134,7 @@ void text_block(const char *str, int max, vector<char *> &lines)
 {
     if(!str) return;
     int visible;
-    while((visible = max ? text_visible(str, max) : strlen(str)))
+    while((visible = max ? text_visible(str, max) : (int)strlen(str)))
     {
         const char *newline = (const char *)memchr(str, '\n', visible);
         if(newline) visible = newline+1-str;
