+++ ac/source/src/tools.cpp
@@ -61,7 +61,7 @@ bool fileexists(const char *path, const char *mode)
 
 bool createdir(const char *path)
 {
-    int len = strlen(path);
+    size_t len = strlen(path);
     if(path[len-1]==PATHDIV)
     {
         static string strip;
@@ -77,7 +77,7 @@ bool createdir(const char *path)
 static void fixdir(char *dir)
 {
     path(dir);
-    int len = strlen(dir);
+    size_t len = strlen(dir);
     if(dir[len-1]!=PATHDIV)
     {
         dir[len] = PATHDIV;
